Add files via upload
Not finished yet. (Tried a lot with the Spring Boot Restful Client with RestTemplate Example but it is not an API)

Now with the Spring Boot 2 REST API Example it works fine but there are still bugs.

Because for the post we need an ID and we have not set one.
And for the get I don't know how to implement the already existing entities.
That's why I added a new class covid to get all values from the covid.json but don't know to realize that exactly.